+++ src/main/kotlin/com/github/michaelbull/advent/day7/Part2.kt
@@ -7,28 +7,28 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
 
-suspend fun Intcode.part2(): Int {
+suspend fun Intcode.part2(): Long {
     return highestThrusterSignal(
         phaseSettings = 5..9,
         getSignal = { feedbackLoopThrusterSignal(it) }
     )
 }
 
-suspend fun Intcode.feedbackLoopThrusterSignal(amplifiers: List<Amplifier>): Int {
+suspend fun Intcode.feedbackLoopThrusterSignal(amplifiers: List<Amplifier>): Long {
     val computers = amplifiers.associateWith { IntcodeComputer() }
 
     return coroutineScope {
         for (i in 0 until amplifiers.size - 1) {
             val curr = amplifiers[i]
             val next = amplifiers[i + 1]
-            val channel = Channel<Int>()
+            val channel = Channel<Long>()
             computers.pair(channel, curr, next)
             addInputs(channel, next)
         }
 
         val first = amplifiers.first()
         val last = amplifiers.last()
-        val feedbackChannel = Channel<Int>(capacity = 1)
+        val feedbackChannel = Channel<Long>(capacity = 1)
         computers.pair(feedbackChannel, last, first)
         addInputs(feedbackChannel, first)
 
@@ -40,9 +40,9 @@ suspend fun Intcode.feedbackLoopThrusterSignal(amplifiers: List<Amplifier>): Int
     }.receive()
 }
 
-private fun CoroutineScope.addInputs(channel: Channel<Int>, amplifier: Amplifier) = launch {
+private fun CoroutineScope.addInputs(channel: Channel<Long>, amplifier: Amplifier) = launch {
     /* Provide each amplifier its phase setting at its first input instruction */
-    channel.send(amplifier.phaseSetting)
+    channel.send(amplifier.phaseSetting.toLong())
 
     /* To start the process, a 0 signal is sent to amplifier A's input exactly once. */
     if (amplifier.id == 'A') {
@@ -56,7 +56,7 @@ private fun CoroutineScope.compute(computer: IntcodeComputer, program: Intcode) 
 }
 
 private suspend fun Map<Amplifier, IntcodeComputer>.pair(
-    channel: Channel<Int>,
+    channel: Channel<Long>,
     curr: Amplifier,
     next: Amplifier
 ) {
